Bug fixes (Use of int instead of size_t for some variables)
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@959 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ model/WithinHost/WithinHostModel.cpp
@@ -46,8 +46,8 @@ double WithinHostModel::immPenalty_22;
 double WithinHostModel::asexImmRemain;
 double WithinHostModel::immEffectorRemain;
 double WithinHostModel::detectionLimit;
-size_t WithinHostModel::totalInfections;
-size_t WithinHostModel::allowedInfections;
+int WithinHostModel::totalInfections;
+int WithinHostModel::allowedInfections;
 
 // -----  static functions  -----
